+++ tests/testthat/test-palaeorotate.R
@@ -23,6 +23,34 @@ test_that("arg 'occdf' works", {
   )
 })
 
+test_that("piping and not piping the first argument give the same result", {
+  skip_if_offline(host = "gws.gplates.org")
+  skip_if_not_installed("vcr")
+
+  occdf <- data.frame(
+    lng = c(2, -103, -66),
+    lat = c(46, 35, -7),
+    age = c(88, 125, 300)
+  )
+
+  vcr::use_cassette("palaeorotate-paleomap", {
+    piped <- occdf |> palaeorotate(model = "PALEOMAP")
+  })
+  vcr::use_cassette("palaeorotate-paleomap", {
+    not_piped <- palaeorotate(occdf, model = "PALEOMAP")
+  })
+
+  expect_equal(piped, not_piped)
+})
+
+test_that("palaeorotate errors with unnamed args", {
+  occdf <- data.frame(lng = c(2, -103), lat = c(46, 35), age = c(88, 125))
+  expect_snapshot(palaeorotate(occdf, "lng"), error = TRUE)
+  expect_snapshot(palaeorotate(occdf = occdf, "lng"), error = TRUE)
+  expect_snapshot(palaeorotate(occdf, "lng", "lat"), error = TRUE)
+  expect_snapshot(palaeorotate(occdf, "lng", lat = "lat"), error = TRUE)
+})
+
 test_that("Large occdf inputs are chunked before being sent to the API", {
   skip_if_offline(host = "gws.gplates.org")
   skip_if_not_installed("vcr")
@@ -376,7 +404,28 @@ test_that("arg 'round' works", {
   expect_snapshot(palaeorotate(occdf = occdf, round = TRUE), error = TRUE)
   expect_snapshot(palaeorotate(occdf = occdf, round = NA), error = TRUE)
   expect_snapshot(palaeorotate(occdf = occdf, round = numeric(0)), error = TRUE)
-
-  # TODO: this should error
   expect_snapshot(palaeorotate(occdf = occdf, round = 1:2), error = TRUE)
+})
+
+test_that("good error message if GPlates or Zenodo are not available", {
+  # We define a "mocked" version of nslookup() that errors on purpose, since
+  # nslookup() would fail if the website is not available.
+  local_mocked_bindings(
+    nslookup = function(...) stop("foo")
+  )
+  occdf <- data.frame(
+    lng = c(2, -103, -66),
+    lat = c(46, 35, -7),
+    age = c(88, 125, 300)
+  )
+  # GPlates
+  expect_snapshot(
+    palaeorotate(occdf = occdf, model = "PALEOMAP"),
+    error = TRUE
+  )
+  # Zenodo
+  expect_snapshot(
+    palaeorotate(occdf = occdf, model = "PALEOMAP", method = "grid"),
+    error = TRUE
+  )
 })